split utils; auto-convert data to base64
