+++ mostlyai/client/model.py
@@ -1,5 +1,5 @@
 # generated by datamodel-codegen:
-#   timestamp: 2024-12-09T11:40:25+00:00
+#   timestamp: 2024-12-10T10:18:57+00:00
 
 from __future__ import annotations
 
@@ -8,9 +8,10 @@ from pathlib import Path
 from typing import Annotated, Any, ClassVar, Dict, List, Literal, Optional, Union
 
 import pandas as pd
-from pydantic import AwareDatetime, Field, RootModel
+from pydantic import AwareDatetime, Field, RootModel, field_validator
 
 from mostlyai.client.base import CustomBaseModel
+from mostlyai.client.base_utils import _convert_to_base64
 
 
 class PermissionLevel(str, Enum):
@@ -1356,6 +1357,20 @@ class SyntheticTableConfiguration(CustomBaseModel):
     tabular_compute: Optional[str] = Field(None, alias="tabularCompute")
     language_compute: Optional[str] = Field(None, alias="languageCompute")
 
+    @field_validator("sample_seed_dict", mode="before")
+    @classmethod
+    def validate_dict_before(cls, value):
+        return (
+            _convert_to_base64(value, format="jsonl")
+            if isinstance(value, dict)
+            else value
+        )
+
+    @field_validator("sample_seed_data", mode="before")
+    @classmethod
+    def validate_data_before(cls, value):
+        return _convert_to_base64(value) if isinstance(value, pd.DataFrame) else value
+
 
 class SyntheticTable(CustomBaseModel):
     id: Optional[str] = Field(
@@ -1552,6 +1567,11 @@ class SourceTableConfig(CustomBaseModel):
     columns: Optional[List[SourceColumnConfig]] = Field(
         None, description="The column configurations of this table."
     )
+
+    @field_validator("data", mode="before")
+    @classmethod
+    def validate_data_before(cls, value):
+        return _convert_to_base64(value) if isinstance(value, pd.DataFrame) else value
 
 
 class SourceTablePatchConfig(CustomBaseModel):
