+++ tools/model.py
@@ -2,8 +2,9 @@ from pathlib import Path
 from typing import Annotated, Any, ClassVar, Literal, Optional, Union
 
 import pandas as pd
-from pydantic import Field
+from pydantic import Field, field_validator
 
+from mostlyai.client.base_utils import _convert_to_base64
 from mostlyai.client.model import (
     JobProgress,
     SyntheticDatasetFormat,
@@ -211,6 +212,29 @@ class Generator:
             raise "Not implemented yet."
             # return self.generator.client._list_synthetic_datasets(self.generator.id)
             pass
+
+
+class SourceTableConfig:
+    @field_validator("data", mode="before")
+    @classmethod
+    def validate_data_before(cls, value):
+        return _convert_to_base64(value) if isinstance(value, pd.DataFrame) else value
+
+
+class SyntheticTableConfiguration:
+    @field_validator("sample_seed_dict", mode="before")
+    @classmethod
+    def validate_dict_before(cls, value):
+        return (
+            _convert_to_base64(value, format="jsonl")
+            if isinstance(value, dict)
+            else value
+        )
+
+    @field_validator("sample_seed_data", mode="before")
+    @classmethod
+    def validate_data_before(cls, value):
+        return _convert_to_base64(value) if isinstance(value, pd.DataFrame) else value
 
 
 class SyntheticDataset:
