+++ tools/postproc_model.py
@@ -25,7 +25,10 @@ def postprocess_model_file(file_path):
             pass
         # Skip the import line for UUID
         elif "import UUID" in line:
-            new_lines.append("import pandas as pd\nfrom pathlib import Path")
+            new_lines.append(
+                "import pandas as pd\nfrom pathlib import Path\n"
+                "from pydantic import field_validator\nfrom mostlyai.client.base_utils import _convert_to_base64"
+            )
         elif "from typing" in line and not import_typing_updated:
             # Append ', ClassVar' to the line if it doesn't already contain ClassVar
             if "ClassVar" not in line:
